requirements.txt
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.2
tabulate>=0.9.0
pytest>=7.4.0
